Rename ImageData to Surface, enable unsafe blocks
