+++ OpenChart/OpenChart.csproj.template
@@ -5,6 +5,7 @@
     <PlatformTarget>x64</PlatformTarget>
     <RuntimeIdentifiers>win-x64;osx-x64;linux-x64</RuntimeIdentifiers>
     <Version>$VERSION</Version>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
     <DebugType>None</DebugType>
